Format code Reorganize code into feature modules Create feature routing modules Create shared module Create modules folder Create interfaces folders Create data folder Create services folder
