+++ APM/src/app/app.component.html
@@ -1,0 +1,12 @@
+<nav class="navbar navbar-expand navbar-light bg-light">
+  <a class="navbar-brand">{{ pageTitle }}</a>
+  <ul class="navbar-nav">
+    <li class="nav-item"><a class="nav-link" routerLinkActive="active" [routerLink]="['/welcome']">Home</a></li>
+    <li class="nav-item"><a class="nav-link" routerLinkActive="active" [routerLinkActiveOptions]="{exact: true}" [routerLink]="['/products']">Product List</a></li>
+    <li class="nav-item"><a class="nav-link" routerLinkActive="active" [routerLinkActiveOptions]="{exact: true}" [routerLink]="['/products/0/edit']">Add Product</a></li>
+  </ul>
+</nav>
+
+<div class="container">
+  <router-outlet></router-outlet>
+</div>
