+++ APM/src/app/app.component.ts
@@ -1,28 +1,11 @@
+// components
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'pm-root',
-  template: `
-    <nav class='navbar navbar-expand navbar-light bg-light'>
-      <a class='navbar-brand'>{{pageTitle}}</a>
-      <ul class='navbar-nav'>
-        <li class='nav-item'><a class='nav-link' routerLinkActive='active'
-              [routerLink]="['/welcome']">Home</a>
-        </li>
-        <li class='nav-item'><a class='nav-link' routerLinkActive='active' [routerLinkActiveOptions]="{exact: true}"
-              [routerLink]="['/products']">Product List</a>
-        </li>
-        <li class='nav-item'><a class='nav-link' routerLinkActive='active' [routerLinkActiveOptions]="{exact: true}"
-              [routerLink]="['/products/0/edit']">Add Product</a>
-        </li>
-      </ul>
-    </nav>
-    <div class='container'>
-      <router-outlet></router-outlet>
-    </div>
-    `,
-  styleUrls: ['./app.component.css']
+  selector: 'app-root',
+  templateUrl: './app.component.html',
+  styleUrls: ['./app.component.scss']
 })
 export class AppComponent {
-  pageTitle = 'Acme Product Management';
+  pageTitle = 'ACME Product Management';
 }
