--- APM/src/app/home/welcome.component.ts
@@ -1,8 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  templateUrl: './welcome.component.html'
-})
-export class WelcomeComponent {
-  public pageTitle = 'Welcome';
-}
