+++ APM/src/app/interfaces/product.ts
@@ -1,0 +1,11 @@
+export interface Product {
+  id: number;
+  productName: string;
+  productCode: string;
+  tags?: string[];
+  releaseDate: string;
+  price: number;
+  description: string;
+  starRating: number;
+  imageUrl: string;
+}
