+++ APM/src/app/modules/home/components/welcome/welcome.component.html
@@ -1,0 +1,21 @@
+﻿<div class="card">
+  <div class="card-header">
+    {{ pageTitle }}
+  </div><!-- /.card-header -->
+
+  <div class="card-body">
+    <div class="container-fluid">
+      <div class="text-center">
+        <img src="./assets/images/logo.jpg" class="img-responsive center-block" style="max-height: 300px; padding-bottom: 50px">
+      </div><!-- /.text-center -->
+
+      <div class="text-center">
+        Developed by:
+      </div><!-- /.text-center -->
+
+      <div class="text-center">
+        <h3>Brook Hutchinson</h3>
+      </div><!-- /.text-center-->
+    </div><!-- /.container-fluid -->
+  </div><!-- /.card-body -->
+</div><!-- /.card -->
