+++ APM/src/app/modules/home/components/welcome/welcome.component.ts
@@ -1,0 +1,9 @@
+// components
+import { Component } from '@angular/core';
+
+@Component({
+  templateUrl: './welcome.component.html'
+})
+export class WelcomeComponent {
+  public pageTitle = 'Welcome';
+}
