+++ APM/src/app/modules/product/components/product-detail/product-detail.component.html
@@ -1,0 +1,63 @@
+<div class="card">
+  <div class="card-header" *ngIf="product">
+    {{ pageTitle + ": " + product.productName }}
+  </div><!-- /.card-header -->
+
+  <div class="card-body" *ngIf="product">
+    <div class="row">
+      <div class="col-md-8">
+        <div class="row">
+          <div class="col-md-3">Name:</div>
+          <div class="col-md-6">{{ product.productName }}</div>
+        </div><!-- /.row -->
+
+        <div class="row">
+          <div class="col-md-3">Code:</div>
+          <div class="col-md-6">{{ product.productCode }}</div>
+        </div><!-- /.row -->
+
+        <div class="row">
+          <div class="col-md-3">Description:</div>
+          <div class="col-md-6">{{ product.description }}</div>
+        </div><!-- /.row -->
+
+        <div class="row">
+          <div class="col-md-3">Availability:</div>
+          <div class="col-md-6">{{ product.releaseDate }}</div>
+        </div><!-- /.row -->
+
+        <div class="row">
+          <div class="col-md-3">Price:</div>
+          <div class="col-md-6">{{ product.price|currency:"USD":"symbol" }}</div>
+        </div><!-- /.row -->
+
+        <div class="row">
+          <div class="col-md-3">5 Star Rating:</div>
+          <div class="col-md-6">
+            <star [rating]="product.starRating"></star>
+          </div>
+        </div><!-- /.row -->
+
+        <div class="row">
+          <div class="col-md-3">Tags:</div>
+          <div class="col-md-6">{{ product.tags }}</div>
+        </div><!-- /.row -->
+      </div>
+
+      <div class="col-md-4" *ngIf="product.imageUrl">
+        <img class="center-block img-responsive" [style.height.px]="150" [style.margin.px]="2" [src]="product.imageUrl" [title]="product.productName">
+      </div>
+    </div><!-- /.row -->
+
+    <div class="row mt-4">
+      <div class="col-md-4">
+        <button class="btn btn-outline-secondary mr-3" style="width: 80px;" (click)="onBack()"><i class="fa fa-chevron-left"></i> Back</button>
+        <button class="btn btn-outline-primary" style="width: 80px;" [routerLink]="['/products', product.id, 'edit']">Edit</button>
+      </div>
+    </div><!-- /row mt-4 -->
+  </div><!-- /.card-body -->
+
+  <div class="alert alert-danger" *ngIf="errorMessage">
+    {{ errorMessage }}
+  </div><!-- /.alert alert-danger -->
+</div><!-- /.card -->
