+++ APM/src/app/modules/product/components/product-list/product-list.component.scss
@@ -1,0 +1,3 @@
+thead {
+  color: #337AB7;
+}
