+++ APM/src/app/modules/product/components/product-list/product-list.component.ts
@@ -1,0 +1,62 @@
+// components
+import { Component, OnInit } from '@angular/core';
+
+// services
+import { ProductService }    from '../../../../services/product.service';
+
+// interfaces
+import { Product }           from '../../../../interfaces/product';
+
+@Component({
+  templateUrl: './product-list.component.html',
+  styleUrls: ['./product-list.component.scss']
+})
+export class ProductListComponent implements OnInit {
+  pageTitle = 'Product List';
+  imageWidth = 50;
+  imageMargin = 2;
+  showImage = false;
+  errorMessage = '';
+
+  _listFilter = '';
+  get listFilter(): string {
+    return this._listFilter;
+  }
+  set listFilter(value: string) {
+    this._listFilter = value;
+    this.filteredProducts = this.listFilter ? this.performFilter(this.listFilter) : this.products;
+  }
+
+  filteredProducts: Product[] = [];
+  products: Product[] = [];
+
+  constructor(private productService: ProductService) {}
+
+  performFilter(filterBy: string): Product[] {
+    filterBy = filterBy.toLocaleLowerCase();
+    return this.products.filter((product: Product) =>
+      product.productName.toLocaleLowerCase().indexOf(filterBy) !== -1);
+  }
+
+  // checks both the product name and tags
+  performFilter2(filterBy: string): Product[] {
+    filterBy = filterBy.toLocaleLowerCase();
+    return this.products.filter((product: Product) =>
+      product.productName.toLocaleLowerCase().indexOf(filterBy) !== -1 ||
+        (product.tags && product.tags.some(tag => tag.toLocaleLowerCase().indexOf(filterBy) !== -1)));
+  }
+
+  toggleImage() {
+    this.showImage = !this.showImage;
+  }
+
+  ngOnInit() {
+    this.productService.getProducts().subscribe({
+      next: products => {
+        this.products = products;
+        this.filteredProducts = this.products;
+      },
+      error: err => this.errorMessage = err
+    });
+  }
+}
