+++ APM/src/app/modules/shared/components/star/star.component.scss
