+++ APM/src/app/modules/shared/components/star/star.component.ts
@@ -1,0 +1,23 @@
+// components
+import { Component, OnChanges } from '@angular/core';
+import { Input, EventEmitter, Output } from '@angular/core';
+
+@Component({
+  selector: 'star',
+  templateUrl: './star.component.html',
+  styleUrls: ['./star.component.scss']
+})
+export class StarComponent implements OnChanges {
+  @Input() rating = 0;
+  @Output() ratingClicked: EventEmitter<string> = new EventEmitter<string>();
+
+  starWidth = 0;
+
+  ngOnChanges(): void {
+    this.starWidth = this.rating * 75 / 5;
+  }
+
+  onClick(): void {
+    this.ratingClicked.emit(`The rating ${this.rating} was clicked!`);
+  }
+}
