+++ APM/src/app/modules/shared/shared.module.spec.ts
@@ -1,0 +1,14 @@
+// modules
+import { SharedModule } from './shared.module';
+
+describe('SharedModule', () => {
+  let sharedModule: SharedModule;
+
+  beforeEach(() => {
+    sharedModule = new SharedModule();
+  });
+
+  it('should create SharedModule', () => {
+    expect(sharedModule).toBeTruthy();
+  });
+});
