+++ APM/src/app/modules/shared/validators/number.validator.ts
@@ -1,0 +1,12 @@
+import { AbstractControl, ValidatorFn } from '@angular/forms';
+
+export class NumberValidators {
+  static range(min: number, max: number): ValidatorFn {
+    return (c: AbstractControl): { [key: string]: boolean } | null => {
+      if (c.value && (isNaN(c.value) || c.value < min || c.value > max)) {
+        return { range: true };
+      }
+      return null;
+    };
+  }
+}
