--- APM/src/app/products/product-detail.component.html
@@ -1,76 +1,0 @@
-<div class="card">
-  <div class="card-header"
-       *ngIf="product">
-    {{pageTitle + ": " + product.productName}}
-  </div>
-
-  <div class="card-body"
-       *ngIf="product">
-
-    <div class="row">
-
-      <div class="col-md-8">
-        <div class="row">
-          <div class="col-md-3">Name:</div>
-          <div class="col-md-6">{{product.productName}}</div>
-        </div>
-        <div class="row">
-          <div class="col-md-3">Code:</div>
-          <div class="col-md-6">{{product.productCode}}</div>
-        </div>
-        <div class="row">
-          <div class="col-md-3">Description:</div>
-          <div class="col-md-6">{{product.description}}</div>
-        </div>
-        <div class="row">
-          <div class="col-md-3">Availability:</div>
-          <div class="col-md-6">{{product.releaseDate}}</div>
-        </div>
-        <div class="row">
-          <div class="col-md-3">Price:</div>
-          <div class="col-md-6">{{product.price|currency:"USD":"symbol"}}</div>
-        </div>
-        <div class="row">
-          <div class="col-md-3">5 Star Rating:</div>
-          <div class="col-md-6">
-            <pm-star [rating]="product.starRating">
-            </pm-star>
-          </div>
-        </div>
-        <div class="row">
-          <div class="col-md-3">Tags:</div>
-          <div class="col-md-6">{{product.tags}}</div>
-        </div>
-      </div>
-
-      <div class="col-md-4"
-           *ngIf="product.imageUrl">
-        <img class="center-block img-responsive"
-             [style.height.px]="150"
-             [style.margin.px]="2"
-             [src]="product.imageUrl"
-             [title]="product.productName">
-      </div>
-    </div>
-
-    <div class="row mt-4">
-      <div class="col-md-4">
-        <button class="btn btn-outline-secondary mr-3"
-                style="width:80px"
-                (click)="onBack()">
-          <i class="fa fa-chevron-left"></i> Back
-        </button>
-        <button class="btn btn-outline-primary"
-                style="width:80px"
-                [routerLink]="['/products', product.id, 'edit']">
-          Edit
-        </button>
-      </div>
-    </div>
-
-  </div>
-
-  <div class="alert alert-danger"
-       *ngIf="errorMessage">{{errorMessage}}
-  </div>
-</div>
