--- APM/src/app/products/product-edit.component.html
@@ -1,139 +1,0 @@
-<div class="card">
-  <div class="card-header">
-    {{pageTitle}}
-  </div>
-
-  <div class="card-body">
-    <form novalidate
-          (ngSubmit)="saveProduct()"
-          [formGroup]="productForm">
-
-      <div class="form-group row mb-2">
-        <label class="col-md-2 col-form-label"
-               for="productNameId">Product Name</label>
-        <div class="col-md-8">
-          <input class="form-control"
-                 id="productNameId"
-                 type="text"
-                 placeholder="Name (required)"
-                 formControlName="productName"
-                 [ngClass]="{'is-invalid': displayMessage.productName }" />
-          <span class="invalid-feedback">
-            {{displayMessage.productName}}
-          </span>
-        </div>
-      </div>
-
-      <div class="form-group row mb-2">
-        <label class="col-md-2 col-form-label"
-               for="productCodeId">Product Code</label>
-        <div class="col-md-8">
-          <input class="form-control"
-                 id="productCodeId"
-                 type="text"
-                 placeholder="Code (required)"
-                 formControlName="productCode"
-                 [ngClass]="{'is-invalid': displayMessage.productCode}" />
-          <span class="invalid-feedback">
-            {{displayMessage.productCode}}
-          </span>
-        </div>
-      </div>
-
-      <div class="form-group row mb-2">
-        <label class="col-md-2 col-form-label"
-               for="starRatingId">Star Rating (1-5)</label>
-
-        <div class="col-md-8">
-          <input class="form-control"
-                 id="starRatingId"
-                 type="text"
-                 placeholder="Rating (1-5)"
-                 formControlName="starRating"
-                 [ngClass]="{'is-invalid': displayMessage.starRating}" />
-          <span class="invalid-feedback">
-            {{displayMessage.starRating}}
-          </span>
-        </div>
-      </div>
-
-      <div formArrayName="tags">
-        <div class="form-group row mb-2"
-             *ngFor="let tag of tags.controls; let i=index">
-          <label class="col-md-2 col-form-label"
-                 [attr.for]="i">Tag</label>
-
-          <div class="col-md-8">
-            <input class="form-control"
-                   [id]="i"
-                   type="text"
-                   placeholder="Tag"
-                   [formControlName]="i" />
-          </div>
-          <button class="col-md-1 btn btn-outline-warning btn-sm"
-                  type="button"
-                  title="Delete this search tag"
-                  (click)="deleteTag(i)">Delete Tag
-          </button>
-        </div>
-      </div>
-
-      <div class="row">
-        <div class="col-md-2 mb-2">
-          <button class="btn btn-outline-secondary btn-sm"
-                  type="button"
-                  title="Add a search tag"
-                  (click)="addTag()">Add Tag
-          </button>
-        </div>
-      </div>
-      
-      <div class="form-group row mb-2">
-        <label class="col-md-2 col-form-label"
-               for="descriptionId">Description</label>
-
-        <div class="col-md-8">
-          <textarea class="form-control"
-                    id="descriptionId"
-                    placeholder="Description"
-                    rows=3
-                    formControlName="description"
-                    [ngClass]="{'is-invalid': displayMessage.description}"></textarea>
-          <span class="invalid-feedback">
-            {{ displayMessage.description}}
-          </span>
-        </div>
-      </div>
-
-      <div class="form-group row mb-2">
-        <div class="offset-md-2 col-md-4">
-          <button class="btn btn-primary mr-3"
-                  style="width:80px;"
-                  type="submit"
-                  [title]="productForm.valid ? 'Save your entered data' : 'Disabled until the form data is valid'"
-                  [disabled]="!productForm.valid">
-            Save
-          </button>
-          <button class="btn btn-outline-secondary mr-3"
-                  style="width:80px;"
-                  type="button"
-                  title="Cancel your edits"
-                  [routerLink]="['/products']">
-            Cancel
-          </button>
-          <button class="btn btn-outline-warning"
-                  style="width:80px"
-                  type="button"
-                  title="Delete this product"
-                  (click)="deleteProduct()">
-            Delete
-          </button>
-        </div>
-      </div>
-    </form>
-  </div>
-
-  <div class="alert alert-danger"
-       *ngIf="errorMessage">{{errorMessage}}
-  </div>
-</div>
