--- APM/src/app/products/product-edit.guard.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed, inject, waitForAsync } from '@angular/core/testing';
-
-import { ProductEditGuard } from './product-edit.guard';
-
-describe('ProductEditGuard', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [ProductEditGuard]
-    });
-  });
-
-  it('should ...', inject([ProductEditGuard], (guard: ProductEditGuard) => {
-    expect(guard).toBeTruthy();
-  }));
-});
