--- APM/src/app/products/product-list.component.css
@@ -1,3 +1,0 @@
-thead {
-  color: #337AB7;
-}
