--- APM/src/app/shared/number.validator.ts
@@ -1,13 +1,0 @@
-import { AbstractControl, ValidatorFn } from '@angular/forms';
-
-export class NumberValidators {
-
-  static range(min: number, max: number): ValidatorFn {
-    return (c: AbstractControl): { [key: string]: boolean } | null => {
-      if (c.value && (isNaN(c.value) || c.value < min || c.value > max)) {
-        return { range: true };
-      }
-      return null;
-    };
-  }
-}
