--- APM/src/app/shared/shared.module.spec.ts
@@ -1,13 +1,0 @@
-import { SharedModule } from './shared.module';
-
-describe('SharedModule', () => {
-  let sharedModule: SharedModule;
-
-  beforeEach(() => {
-    sharedModule = new SharedModule();
-  });
-
-  it('should create an instance', () => {
-    expect(sharedModule).toBeTruthy();
-  });
-});
