--- APM/src/app/shared/shared.module.ts
@@ -1,20 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
-
-import { StarComponent } from './star.component';
-
-@NgModule({
-  imports: [
-    CommonModule
-  ],
-  declarations: [
-    StarComponent
-  ],
-  exports: [
-    StarComponent,
-    CommonModule,
-    FormsModule
-  ]
-})
-export class SharedModule { }
