--- APM/src/app/shared/star.component.html
@@ -1,12 +1,0 @@
-<div class="crop"
-     [style.width.px]="starWidth"
-     [title]="rating"
-     (click)="onClick()">
-  <div style="width: 75px">
-    <span class="fa fa-star"></span>
-    <span class="fa fa-star"></span>
-    <span class="fa fa-star"></span>
-    <span class="fa fa-star"></span>
-    <span class="fa fa-star"></span>
-  </div>
-</div>
